Migrate old Gauntlet code to GauntletLegacy target.
