+++ Package.swift
@@ -14,7 +14,8 @@ let package = Package(
     ],
     products: [
         // Products define the executables and libraries produced by a package, and make them visible to other packages.
-        .library(name: "Gauntlet", targets: ["Gauntlet"])
+        .library(name: "Gauntlet", targets: ["Gauntlet"]),
+        .library(name: "GauntletLegacy", targets: ["GauntletLegacy"])
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
@@ -24,6 +25,8 @@ let package = Package(
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages which this package depends on.
         .target(name: "Gauntlet", dependencies: []),
-        .testTarget(name: "GauntletTests", dependencies: ["Gauntlet"])
+        .target(name: "GauntletLegacy", dependencies: []),
+        .testTarget(name: "GauntletTests", dependencies: ["Gauntlet"]),
+        .testTarget(name: "GauntletLegacyTests", dependencies: ["GauntletLegacy"])
     ]
 )
