+++ Source/Gauntlet/Documentation.docc/Documentation.md
@@ -60,40 +60,3 @@ These types allow explicitly handling throwing code in assertions.
 These are types used in unit tests for Gauntlet, and can be used for testing custom assertion operators.
 - ``MockFailureRecorder``
 - ``RecordedFailure``
-
-### Legacy
-
-These are legacy functions and types from the original API. These still exist for backwards compatibility, but it is recommend you move to the newer API as these will be deprected and removed in future releases.
-
-- ``XCTAwaitAssertEqual(_:_:_:reporter:file:line:then:)-9l6yg``
-- ``XCTAwaitAssertEqual(_:_:_:reporter:file:line:then:)-5tofs``
-- ``XCTAwaitAssertFalse(_:_:reporter:file:line:then:)``
-- ``XCTAwaitAssertNoThrow(_:_:reporter:file:line:then:)``
-- ``XCTAwaitAssertNotNil(_:_:reporter:file:line:then:)``
-- ``XCTAwaitAssertThrowsError(_:equalTo:_:reporter:file:line:then:)``
-- ``XCTAwaitAssertThrowsError(_:ofType:_:reporter:file:line:then:)``
-- ``XCTAwaitAssertTrue(_:_:reporter:file:line:then:)``
-- ``XCTAssertTrue(_:_:reporter:file:line:then:)``
-- ``XCTAssertFalse(_:_:reporter:file:line:then:)``
-- ``XCTAssertIsEmpty(_:_:reporter:file:line:then:)``
-- ``XCTAssertIsNotEmpty(_:_:reporter:file:line:then:)``
-- ``XCTAssertOnQueue(_:_:reporter:file:line:then:)``
-- ``XCTAssertEqual(_:_:_:reporter:file:line:then:)``
-- ``XCTAssertIdentical(_:_:_:reporter:file:line:then:)``
-- ``XCTAssertNotIdentical(_:_:_:reporter:file:line:then:)``
-- ``XCTAssertZero(_:_:reporter:file:line:then:)``
-- ``XCTAssertNotNil(_:_:reporter:file:line:then:)``
-- ``XCTAssertFailure(_:_:reporter:file:line:then:)``
-- ``XCTAssertFailure(_:equalTo:_:reporter:file:line:then:)``
-- ``XCTAssertFailure(_:is:_:reporter:file:line:then:)``
-- ``XCTAssertFailure(_:is:equalTo:_:reporter:file:line:then:)``
-- ``XCTAssertSuccess(_:_:reporter:file:line:then:)``
-- ``XCTAssertSuccess(_:is:_:reporter:file:line:then:)``
-- ``XCTAssertSuccess(_:equalTo:_:reporter:file:line:then:)``
-- ``XCTAssertSuccess(_:is:equalTo:_:reporter:file:line:then:)``
-- ``XCTAssertSuccess(_:keyPath:isEqualTo:_:reporter:file:line:then:)``
-- ``XCTAssert(_:contains:_:reporter:file:line:then:)``
-- ``XCTAssert(_:is:_:reporter:file:line:then:)``
-- ``XCTAssert(_:is:equalTo:_:reporter:file:line:then:)``
-- ``XCTestReporter``
-- ``FailureReporter``
