+++ Tests/GauntletLegacyTests/Utility/EquatableClass.swift
