+++ Tests/GauntletLegacyTests/XCTAsserts/XCTCollectionAssertsTests.swift
@@ -1,4 +1,4 @@
-//  XCTTypeAssertsTests.swift
+//  XCTCollectionAssertsTests.swift
 //
 //  MIT License
 //
@@ -22,42 +22,40 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
-import Gauntlet
-import Foundation
+import GauntletLegacy
 import XCTest
 
-class XCTTypeAssertsTestCase: XCTestCase {
+// swiftlint:disable discouraged_optional_collection
 
-    // MARK: - Assert(_, is:)
+class XCTCollectionAssertsTestCase: XCTestCase {
 
-    func testTypeAssert() {
+    // MARK: - XCTAssertIsEmpty
+
+    func testAssertIsEmpty() {
         // Given
+        let array: [Int] = []
         let mock = FailMock()
-        let anyInt: Any = 5
         var completionCalled = false
-        var valuePassedToCompletion: Int?
 
         // When
-        XCTAssert(anyInt, is: Int.self, reporter: mock) { value in
+        XCTAssertIsEmpty(array, "custom message", reporter: mock, file: "some file", line: 123) {
             completionCalled = true
-            valuePassedToCompletion = value
         }
 
         // Then
         XCTAssertTrue(completionCalled)
-        XCTAssertEqual(valuePassedToCompletion, 5)
         XCTAssertNil(mock.message)
         XCTAssertNil(mock.file)
         XCTAssertNil(mock.line)
     }
 
-    func testTypeAssertNoClosure() {
+    func testAssertIsEmptyNoClosure() {
         // Given
+        let array: [Int] = []
         let mock = FailMock()
-        let anyInt: Any = 5
 
         // When
-        XCTAssert(anyInt, is: Int.self, reporter: mock)
+        XCTAssertIsEmpty(array, "custom message", reporter: mock, file: "some file", line: 123)
 
         // Then
         XCTAssertNil(mock.message)
@@ -65,92 +63,70 @@ class XCTTypeAssertsTestCase: XCTestCase {
         XCTAssertNil(mock.line)
     }
 
-    func testFailingTypeAssert() {
+    func testFailingAssertIsEmpty() {
         // Given
+        let array = [1, 2, 3]
         let mock = FailMock()
         var completionCalled = false
 
         // When
-        XCTAssert(5, is: String.self, "custom message", reporter: mock, file: "some file", line: 123) { _ in
+        XCTAssertIsEmpty(array, "custom message", reporter: mock, file: "some file", line: 123) {
             completionCalled = true
         }
 
         // Then
         XCTAssertFalse(completionCalled)
-        XCTAssertEqual(mock.message, "XCTAssert(_, is:) - value of type Int is not expected type String - custom message")
+        XCTAssertEqual(mock.message, "XCTAssertIsEmpty - custom message")
         XCTAssertEqual(mock.file, "some file")
         XCTAssertEqual(mock.line, 123)
     }
 
-    func testFailingTypeAssertOnNil() {
+    func testThrowingAssertIsEmpty() {
         // Given
         let mock = FailMock()
+        let expression: () throws -> [Int] = { throw MockError.someError }
         var completionCalled = false
 
         // When
-        XCTAssert(nil, is: Int.self, "custom message", reporter: mock, file: "some file", line: 123) { _ in
+        XCTAssertIsEmpty(try expression(), "custom message", reporter: mock, file: "some file", line: 123) {
             completionCalled = true
         }
 
         // Then
         XCTAssertFalse(completionCalled)
-        XCTAssertEqual(
-            mock.message,
-            "XCTAssert(_, is:) - value is nil and cannot be checked for conformance to type Int - custom message"
-        )
-
+        XCTAssertEqual(mock.message, #"XCTAssertIsEmpty - threw error "Some Error" - custom message"#)
         XCTAssertEqual(mock.file, "some file")
         XCTAssertEqual(mock.line, 123)
     }
 
-    func testThrowingTypeAssert() {
+    func testAssertIsEmptyThrowingInThen() {
         // Given
-        let function: () throws -> Any = { throw MockError.someError }
         let mock = FailMock()
+        let expression: () throws -> [Int] = { throw MockError.someError }
         var completionCalled = false
 
         // When
-        XCTAssert(try function(), is: String.self, "custom message", reporter: mock, file: "some file", line: 123) { _ in
+        XCTAssertIsEmpty(try expression(), "custom message", reporter: mock, file: "some file", line: 123) {
             completionCalled = true
         }
 
         // Then
         XCTAssertFalse(completionCalled)
-        XCTAssertEqual(mock.message, #"XCTAssert(_, is:) - threw error "Some Error" - custom message"#)
+        XCTAssertEqual(mock.message, #"XCTAssertIsEmpty - threw error "Some Error" - custom message"#)
         XCTAssertEqual(mock.file, "some file")
         XCTAssertEqual(mock.line, 123)
     }
 
-    func testTypeAssertWithThrowingThen() {
-        // Given
-        let anyString: Any = ""
-        let mock = FailMock()
-        var completionCalled = false
-
-        // When
-        XCTAssert(anyString, is: String.self, "custom message", reporter: mock, file: "some file", line: 123) { _ in
-            completionCalled = true
-            throw MockError.someError
-        }
-
-        // Then
-        XCTAssertTrue(completionCalled)
-        XCTAssertEqual(mock.message, #"XCTAssert(_, is:) - then closure threw error "Some Error" - custom message"#)
-        XCTAssertEqual(mock.file, "some file")
-        XCTAssertEqual(mock.line, 123)
-    }
-
-    func testTypeAssertWithThrowingCallInAutoclosureInThen() {
+    func testAssertIsEmptyWithThrowingCallInAutoclosureInThen() {
         // This is a compile time test that ensures that a throwing function can be called within an autoclosure
         // in the then closure of this assert. This simulates calling throwing code inside another assert within
         // the closure.
 
         // Given
-        let anyInt: Any = 5
+        let emptyArray: [Int] = []
         var completionCalled = false
 
-        // When
-        XCTAssert(anyInt, is: Int.self) { _ in
+        XCTAssertIsEmpty(emptyArray) {
             throwingAutoclosure(try functionThatCanThrow())
             completionCalled = true
         }
@@ -159,33 +135,36 @@ class XCTTypeAssertsTestCase: XCTestCase {
         XCTAssertTrue(completionCalled)
     }
 
-    // MARK: - Assert(_, is:, equalTo:)
+    // MARK: - XCTAssertIsNotEmpty
 
-    func testTypeEqualityAssert() {
+    func testAssertIsNotEmpty() {
         // Given
+        let array = [1, 2, 3]
         let mock = FailMock()
-        let any: Any = 5
         var completionCalled = false
+        var arrayPassedToClosure: [Int]?
 
         // When
-        XCTAssert(any, is: Int.self, equalTo: 5, reporter: mock) {
+        XCTAssertIsNotEmpty(array, "custom message", reporter: mock, file: "some file", line: 123) { array in
+            arrayPassedToClosure = array
             completionCalled = true
         }
 
         // Then
         XCTAssertTrue(completionCalled)
+        XCTAssertEqual(arrayPassedToClosure, array)
         XCTAssertNil(mock.message)
         XCTAssertNil(mock.file)
         XCTAssertNil(mock.line)
     }
 
-    func testTypeEqualityAssertNoClosure() {
+    func testAssertIsNotEmptyNoClosure() {
         // Given
+        let array = [1, 2, 3]
         let mock = FailMock()
-        let any: Any = EquatableClass(value: 5)
 
         // When
-        XCTAssert(any, is: EquatableClass.self, equalTo: EquatableClass(value: 5), reporter: mock)
+        XCTAssertIsNotEmpty(array, "custom message", reporter: mock, file: "some file", line: 123)
 
         // Then
         XCTAssertNil(mock.message)
@@ -193,123 +172,87 @@ class XCTTypeAssertsTestCase: XCTestCase {
         XCTAssertNil(mock.line)
     }
 
-    func testFailingTypeEqualityAssertOnUnequalValue() {
+    func testFailingAssertIsNotEmpty() {
         // Given
-        let anyInt: Any = 5
+        let array: [Int] = []
         let mock = FailMock()
         var completionCalled = false
 
         // When
-        XCTAssert(anyInt, is: Int.self, equalTo: 1, "custom message", reporter: mock, file: "some file", line: 123) {
+        XCTAssertIsNotEmpty(array, "custom message", reporter: mock, file: "some file", line: 123) { _ in
             completionCalled = true
         }
 
         // Then
         XCTAssertFalse(completionCalled)
-        XCTAssertEqual(mock.message, "XCTAssert(_, is:, equalTo:) - 5 is not equal to 1 - custom message")
+        XCTAssertEqual(mock.message, "XCTAssertIsNotEmpty - custom message")
         XCTAssertEqual(mock.file, "some file")
         XCTAssertEqual(mock.line, 123)
     }
 
-    func testFailingTypeEqualityAssert() {
+    func testFailingAssertIsNotEmptyWithNil() {
         // Given
+        let array: [Int]? = nil
         let mock = FailMock()
         var completionCalled = false
 
         // When
-        XCTAssert(5, is: String.self, equalTo: "", "custom message", reporter: mock, file: "some file", line: 123) {
+        XCTAssertIsNotEmpty(array, "custom message", reporter: mock, file: "some file", line: 123) { _ in
             completionCalled = true
         }
 
         // Then
         XCTAssertFalse(completionCalled)
-        XCTAssertEqual(
-            mock.message,
-            "XCTAssert(_, is:, equalTo:) - value of type Int is not expected type String - custom message"
-        )
-
+        XCTAssertEqual(mock.message, "XCTAssertIsNotEmpty - custom message")
         XCTAssertEqual(mock.file, "some file")
         XCTAssertEqual(mock.line, 123)
     }
 
-    func testFailingTypeEqualityAssertOnNil() {
+    func testThrowingAssertIsNotEmpty() {
         // Given
+        let expression: () throws -> [Int] = { throw MockError.someError }
         let mock = FailMock()
         var completionCalled = false
 
         // When
-        XCTAssert(nil, is: Int.self, equalTo: 5, "custom message", reporter: mock, file: "some file", line: 123) {
+        XCTAssertIsNotEmpty(try expression(), "custom message", reporter: mock, file: "some file", line: 123) { _ in
             completionCalled = true
         }
 
         // Then
         XCTAssertFalse(completionCalled)
-        XCTAssertEqual(
-            mock.message,
-            "XCTAssert(_, is:, equalTo:) - value is nil and cannot be checked for conformance to type Int - custom message"
-        )
-
+        XCTAssertEqual(mock.message, #"XCTAssertIsNotEmpty - threw error "Some Error" - custom message"#)
         XCTAssertEqual(mock.file, "some file")
         XCTAssertEqual(mock.line, 123)
     }
 
-    func testThrowingTypeEqualityAssert() {
+    func testAssertIsNotEmptyWithThrowingThen() {
         // Given
-        let function: () throws -> Any = { throw MockError.someError }
         let mock = FailMock()
         var completionCalled = false
 
         // When
-        XCTAssert(
-            try function(),
-            is: String.self,
-            equalTo: "",
-            "custom message",
-            reporter: mock,
-            file: "some file",
-            line: 123,
-            then: {
-                completionCalled = true
-            }
-        )
-
-        // Then
-        XCTAssertFalse(completionCalled)
-        XCTAssertEqual(mock.message, "XCTAssert(_, is:, equalTo:) - threw error \"Some Error\" - custom message")
-        XCTAssertEqual(mock.file, "some file")
-        XCTAssertEqual(mock.line, 123)
-    }
-
-    func testTypeEqualityAssertWithThrowingThen() {
-        // Given
-        let anyInt: Any = 5
-        let mock = FailMock()
-        var completionCalled = false
-
-        // When
-        XCTAssert(anyInt, is: Int.self, equalTo: 5, "custom message", reporter: mock, file: "some file", line: 123) {
+        XCTAssertIsNotEmpty([1, 2, 3], "custom message", reporter: mock, file: "some file", line: 123) { _ in
             completionCalled = true
             throw MockError.someError
         }
 
         // Then
-        XCTAssertTrue(completionCalled)
-        XCTAssertEqual(mock.message, #"XCTAssert(_, is:, equalTo:) - then closure threw error "Some Error" - custom message"#)
+        XCTAssert(completionCalled)
+        XCTAssertEqual(mock.message, #"XCTAssertIsNotEmpty - then closure threw error "Some Error" - custom message"#)
         XCTAssertEqual(mock.file, "some file")
         XCTAssertEqual(mock.line, 123)
     }
 
-    func testTypeEqualtyAssertWithThrowingCallInAutoclosureInThen() {
+    func testAssertIsNotEmptyWithThrowingCallInAutoclosureInThen() {
         // This is a compile time test that ensures that a throwing function can be called within an autoclosure
         // in the then closure of this assert. This simulates calling throwing code inside another assert within
         // the closure.
 
         // Given
-        let anyInt: Any = 5
         var completionCalled = false
 
-        // When
-        XCTAssert(anyInt, is: Int.self, equalTo: 5) {
+        XCTAssertIsNotEmpty([1, 2, 3]) { _ in
             throwingAutoclosure(try functionThatCanThrow())
             completionCalled = true
         }
