+++ Tests/GauntletLegacyTests/XCTAsserts/XCTEqualityAssertsTests.swift
@@ -22,7 +22,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
-import Gauntlet
+import GauntletLegacy
 import XCTest
 
 // swiftlint:disable type_body_length file_length
